% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naviguer.R
\name{lister_exercices}
\alias{lister_exercices}
\title{Lister les exercices du tutoriel}
\usage{
lister_exercices()
}
\value{
Invisiblement, les codes des exercices (\code{character}).
}
\description{
Affiche les exercices disponibles avec, pour chacun, son intention. Sert à
s'orienter ; ne dévoile pas les étapes de résolution.
}
\examples{
lister_exercices()
}
