% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer-projet.R
\name{comparer_chartes}
\alias{comparer_chartes}
\title{Comparer les couleurs des variantes de charte}
\usage{
comparer_chartes()
}
\value{
Invisiblement, un vecteur nommé des couleurs principales.
}
\description{
Affiche la couleur principale (\code{primary}) de chaque variante de charte, avec
un aperçu coloré (si votre console le supporte).
}
\examples{
comparer_chartes()
}
