% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutoquartotypst-package.R
\docType{package}
\name{tutoquartotypst-package}
\alias{tutoquartotypst}
\alias{tutoquartotypst-package}
\title{tutoquartotypst: Paquet Compagnon du Tutoriel Quarto + Typst (Rencontres R 2026)}
\description{
Fonctions utilitaires pour preparer et suivre le tutoriel "PDF sans frictions : Typst dans vos projets Quarto" (Rencontres R 2026). Le paquet verifie l'environnement (R, Quarto, Typst, paquets), installe localement les exercices et aide au rendu Typst. Installer le paquet tire automatiquement tous les prerequis du tutoriel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cderv/tuto-quarto-typst-rr-2026}
  \item \url{https://connect.posit.cloud/cderv/content/019df82c-7202-2165-28ea-7f9ca734ad26/package}
  \item Report bugs at \url{https://github.com/cderv/tuto-quarto-typst-rr-2026/issues}
}

}
\author{
\strong{Maintainer}: Christophe Dervieux \email{christophe.dervieux@gmail.com}

Authors:
\itemize{
  \item Christophe Dervieux \email{christophe.dervieux@gmail.com}
}

}
\keyword{internal}
